Creation of selection sets fixed and more robust
Changed the creation of the selection sets.
Now the uv values are evaluated via the polygon and not the vertices.
This makes sure that only the poly in a specific quadrant is selected.
The previous method via vertex lead to double selection sometimes. So
polys in other quadrants could end up in different selection sets.
This is now fixed and the progress bar works now as well.